Added "disable" functionality.
Now when you click the extension icon, it not only removes the purple
border from the current page, it also disables the extension (meaning
it won’t run on new pages).

The way the border is removed as also been improved. Previously, if a
link had a border, this extension would overwrite it, and when you
turned off the extension, the link would have no border. Now, when the
extension is turned off, the original border will show.